feat: Move Site Editor navigation command from editor to core-commands
